fix: merge missing changes
