+++ packages/stg-evm-v2/devtools/config/utils/circle-fiat-token.config.utils.ts
@@ -1,59 +1,56 @@
 import { StargateType, TokenName } from '@stargatefinance/stg-definitions-v2'
-import { USDCNodeConfig } from '@stargatefinance/stg-devtools-v2'
+import { CircleFiatTokenNodeConfig } from '@stargatefinance/stg-devtools-v2'
 
 import { OmniGraphHardhat, createContractFactory, createGetHreByEid } from '@layerzerolabs/devtools-evm-hardhat'
 import { Stage } from '@layerzerolabs/lz-definitions'
 
 import { getCircleFiatTokenProxyDeployName } from '../../../ops/util'
-import { createGetAssetAddresses, createGetNamedAccount, getAssetNetworkConfig } from '../../../ts-src/utils/util'
+import { createGetAssetAddresses, getAssetNetworkConfig } from '../../../ts-src/utils/util'
 import { getContractWithEid, getSafeAddress } from '../utils'
 import { getChainsThatSupportTokenWithType, isExternalDeployment, setStage } from '../utils/utils.config'
 
-const tokenName = TokenName.USDC
-const fiatContract = { contractName: 'FiatTokenV2_2' }
-const proxyContract = { contractName: getCircleFiatTokenProxyDeployName(tokenName) }
-
-export default async function buildUsdcTokenGraph(stage: Stage): Promise<OmniGraphHardhat<USDCNodeConfig, unknown>> {
+export default async function buildCircleFiatTokenGraph(
+    stage: Stage,
+    tokenName: TokenName
+): Promise<OmniGraphHardhat<CircleFiatTokenNodeConfig, unknown>> {
     // Set the correct stage
     setStage(stage)
+
+    const proxyContract = { contractName: getCircleFiatTokenProxyDeployName(tokenName) }
+    const fiatContract = { contractName: 'FiatTokenV2_2' }
 
     // First let's create the HardhatRuntimeEnvironment objects for all networks
     const getEnvironment = createGetHreByEid()
     const contractFactory = createContractFactory(getEnvironment)
     const getAssetAddresses = createGetAssetAddresses(getEnvironment)
-    const getStargateMultisigTestnet = createGetNamedAccount(getEnvironment)
 
-    // The newer USDC deployments (since December 2024)
+    // note: The newer USDC deployments (since December 2024, USDC is deployed and verified from Circle's repo)
     const chains = getChainsThatSupportTokenWithType(tokenName, StargateType.Oft)
     const contracts = await Promise.all(
         chains.map(async (chain) => {
-            let usdcProxyAddress
+            let tokenProxyAddress
             if (isExternalDeployment(chain, tokenName)) {
                 // if is external deployment, we need to get the fiat token proxy address
-                usdcProxyAddress = await contractFactory(
+                tokenProxyAddress = await contractFactory(
                     getContractWithEid(chain.eid, {
                         contractName: 'FiatTokenProxy',
-                        address: getAssetNetworkConfig(chain.eid, tokenName).address, // usdc asset address
+                        address: getAssetNetworkConfig(chain.eid, tokenName).address, // eurc/usdc asset address
                     })
                 )
             } else {
-                // if is not external deployment, we need to get USDCProxy address
-                usdcProxyAddress = await contractFactory(getContractWithEid(chain.eid, proxyContract))
+                // if is not external deployment, we need to get tokenProxy address
+                tokenProxyAddress = await contractFactory(getContractWithEid(chain.eid, proxyContract))
             }
 
-            const stargateMultisig =
-                stage === Stage.MAINNET
-                    ? getSafeAddress(chain.eid)
-                    : await getStargateMultisigTestnet(chain.eid, 'usdcAdmin')
+            const stargateMultisig = getSafeAddress(chain.eid)
             const assetAddresses = await getAssetAddresses(chain.eid, [tokenName])
             return {
                 contract: getContractWithEid(chain.eid, {
                     ...fiatContract,
-                    address: usdcProxyAddress.contract.address,
+                    address: tokenProxyAddress.contract.address,
                 }),
                 config: {
-                    // Only set owner for mainnet
-                    ...(stage === Stage.MAINNET ? { owner: getSafeAddress(chain.eid) } : {}),
+                    owner: stargateMultisig,
                     masterMinter: stargateMultisig,
                     pauser: stargateMultisig,
                     rescuer: stargateMultisig,
